backend/requirements.txt
fastapi
uvicorn
python-dotenv
pyjwt
passlib[bcrypt]
requests
